refactor: renderer related codes
