+++ src/raycast/Detail/index.tsx
@@ -1,6 +1,6 @@
 import { Detail as RDetail } from "raycast-original";
 
-import * as ElementTypes from "../../elements/types";
+import * as ElementTypes from "../../renderer/elements/types";
 
 type DetailPropKeys = (keyof RDetail.Props)[];
 const serializedKeys: DetailPropKeys = ["isLoading", "markdown", "navigationTitle"];
