+++ src/raycast/List/EmptyView.tsx
@@ -1,6 +1,6 @@
 import { List } from "raycast-original";
 
-import * as ElementTypes from "../../elements/types";
+import * as ElementTypes from "../../renderer/elements/types";
 
 type EmptyViewPropKeys = (keyof List.EmptyView.Props)[];
 const serializedKeys: EmptyViewPropKeys = ["title", "description", "icon"];
