+++ src/renderer/elements/Command.ts
@@ -1,7 +1,7 @@
 import type { Server } from "rpc-websockets";
 
+import { createDebug } from "../../utils/debug";
 import { Container } from "../types";
-import { createDebug } from "../utils/debug";
 
 import BaseElement from "./BaseElement";
