+++ src/renderer/elements/PrimitiveElements.ts
@@ -1,6 +1,6 @@
 import pick from "lodash/pick";
 
-import { createDebug } from "../utils/debug";
+import { createDebug } from "../../utils/debug";
 
 import BaseElement from "./BaseElement";
