+++ src/renderer/run.ts
@@ -1,9 +1,9 @@
 import React from "react";
 import type { Server } from "rpc-websockets";
 
-import { render } from "../render";
+import { App } from "../examples/App";
 
-import { App } from "./App";
+import { render } from "./render";
 
 export function run(Command: React.FC, { server }: { server: Server }) {
   render(React.createElement(App, { server, Command }), server);
