+++ src/renderer/types.ts
